pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-protobuf"]
build-backend = "setuptools.build_meta"

[project]
name = "gptest"
version = "0.1.0"
authors = [
  {name = "Chris", email = "99300381+ckrbl@users.noreply.github.com"},
]
description = "An exemplar package showing a client and server communicating with GRPC and Protobufs"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = ["click", "grpcio", "grpcio-tools"]

[tool]
[tool.setuptools]
packages = ["gptest.client", "gptest.proto", "gptest.server"]

[tool.setuptools-protobuf]
protobufs = ["gptest/proto/message.proto"]
